// src/lib/sanity/schema-types.ts
export type Settings = {
  _id: string
  _type: 'settings'
  _createdAt: string
  _updatedAt: string
  _rev: string
  brandName?: string
  localizedTaglines?: Array<{
    language?: 'en' | 'fr' | 'es' | 'de'
    tagline?: string
    _type: 'localizedTagline'
    _key: string
  }>
  logo?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  logoAlt?: string
  contactEmail?: string
  favicons?: {
    favicon16?: {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      media?: unknown
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      _type: 'image'
    }
    favicon32?: {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      media?: unknown
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      _type: 'image'
    }
    favicon96?: {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      media?: unknown
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      _type: 'image'
    }
    favicon180?: {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      media?: unknown
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      _type: 'image'
    }
    favicon192?: {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      media?: unknown
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      _type: 'image'
    }
    favicon512?: {
      asset?: {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
      }
      media?: unknown
      hotspot?: SanityImageHotspot
      crop?: SanityImageCrop
      _type: 'image'
    }
  }
  publishedLanguages?: Array<{
    code?: 'en' | 'fr' | 'es' | 'de'
    label?: string
    _type: 'language'
    _key: string
  }>
  localizedMenus?: Array<{
    language?: 'en' | 'fr' | 'es' | 'de'
    menu?: Menu
    _type: 'localizedMenu'
    _key: string
  }>
  menu?: Menu
  localizedFooters?: Array<{
    language?: 'en' | 'fr' | 'es' | 'de'
    footer?: FooterSettings
    _type: 'localizedFooter'
    _key: string
  }>
  footer?: FooterSettings
  customProductOptions?: Array<
    | ({
        _key: string
      } & CustomProductOptionColor)
    | ({
        _key: string
      } & CustomProductOptionSize)
  >
  localizedNotFoundPages?: Array<{
    language?: 'en' | 'fr' | 'es' | 'de'
    notFound?: NotFoundPage
    _type: 'localizedNotFound'
    _key: string
  }>
  notFoundPage?: NotFoundPage
  socialLinks?: Array<{
    platform?:
      | 'instagram'
      | 'facebook'
      | 'twitter'
      | 'tiktok'
      | 'youtube'
      | 'pinterest'
      | 'linkedin'
    url?: string
    _type: 'socialLink'
    _key: string
  }>
  localizedSeo?: Array<{
    language?: 'en' | 'fr' | 'es' | 'de'
    seo?: Seo
    _type: 'localizedSeo'
    _key: string
  }>
  seo?: Seo
}
// ...
export type NotFoundPage = {
  _type: 'notFoundPage'
  title?: string
  body?: string
  buttonLabel?: string
  buttonLink?: Array<
    | ({
        _key: string
      } & LinkInternal)
    | ({
        _key: string
      } & LinkExternal)
  >
}

export type Menu = {
  _type: 'menu'
  links?: Array<
    | ({
        _key: string
      } & CollectionGroup)
    | ({
        _key: string
      } & LinkInternal)
    | ({
        _key: string
      } & LinkExternal)
  >
  theme?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'theme'
  }
}
// ...
export type FooterSettings = {
  _type: 'footerSettings'
  navigations?: Array<{
    _ref: string
    _type: 'reference'
    _weak?: boolean
    _key: string
    [internalGroqTypeReferenceTo]?: 'navigation'
  }>
  legalBar?: {
    navigation?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'navigation'
    }
  }
  theme?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'theme'
  }
}
// ...
export type CustomProductOptionSize = {
  _type: 'customProductOption.size'
  title?: string
  sizes?: Array<
    {
      _key: string
    } & CustomProductOptionSizeObject
  >
}

export type CustomProductOptionSizeObject = {
  _type: 'customProductOption.sizeObject'
  title?: string
  width?: number
  height?: number
}

export type CustomProductOptionColor = {
  _type: 'customProductOption.color'
  title?: string
  colors?: Array<
    {
      _key: string
    } & CustomProductOptionColorObject
  >
}

export type CustomProductOptionColorObject = {
  _type: 'customProductOption.colorObject'
  title?: string
  color?: Color
}
// ...
export type CollectionGroup = {
  _type: 'collectionGroup'
  title?: string
  collectionLinks?: CollectionLinks
  collectionProducts?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'collection'
  }
}
// ...
export type LinkInternal = {
  _type: 'linkInternal'
  label?: string
  reference?:
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'collection'
      }
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'home'
      }
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'page'
      }
    | {
        _ref: string
        _type: 'reference'
        _weak?: boolean
        [internalGroqTypeReferenceTo]?: 'product'
      }
}
// ...
export type LinkExternal = {
  _type: 'linkExternal'
  url?: string
  newWindow?: boolean
}
// ...
export type Seo = {
  _type: 'seo'
  title?: string
  description?: string
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    media?: unknown
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
}
// ...
export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}

export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol